Remove ancient MessageTest duplicate
We have a more up-to-date test in

org.graylog2.plugin.MessageTest
