+++ graylog2-server/src/test/java/org/graylog2/plugin/MessageTest.java
@@ -24,6 +24,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
+import org.graylog.failure.FailureCause;
 import org.graylog2.indexer.IndexSet;
 import org.graylog2.plugin.streams.Stream;
 import org.graylog2.shared.SuppressForbidden;
@@ -55,6 +56,7 @@ import java.util.regex.Pattern;
 
 import static com.google.common.collect.Sets.symmetricDifference;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.graylog2.plugin.streams.Stream.DEFAULT_STREAM_ID;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -668,5 +670,69 @@ public class MessageTest {
         message.setMetadata("stateKey", 10L);
         assertThat(message.getMetadataValue("badKey", "default")).isEqualTo("default");
         assertThat(message.getMetadataValue("stateKey", "default")).isEqualTo(10L);
+    }
+
+    @Test
+    public void addProcessingError_appendsWithEachCall() {
+        final Message msg = new Message(new org.testcontainers.shaded.com.google.common.collect.ImmutableMap.Builder<String, Object>()
+                .put(Message.FIELD_ID, "msg-id")
+                .put(Message.FIELD_TIMESTAMP, Tools.buildElasticSearchTimeFormat(Tools.nowUTC()))
+                .build());
+
+        final FailureCause cause1 = () -> "Cause 1";
+        final FailureCause cause2 = () -> "Cause 2";
+
+        msg.addProcessingError(new Message.ProcessingError(cause1, "Failure Message #1", "Failure Details #1"));
+
+        assertThat(msg.processingErrors())
+                .containsExactly(new Message.ProcessingError(cause1, "Failure Message #1", "Failure Details #1"));
+
+        msg.addProcessingError(new Message.ProcessingError(cause2, "Failure Message #2", "Failure Details #2"));
+
+        assertThat(msg.processingErrors())
+                .containsExactly(
+                        new Message.ProcessingError(cause1, "Failure Message #1", "Failure Details #1"),
+                        new Message.ProcessingError(cause2, "Failure Message #2", "Failure Details #2"));
+    }
+
+    @Test
+    public void processingErrors_returnImmutableList() {
+        final Message msg = new Message(new org.testcontainers.shaded.com.google.common.collect.ImmutableMap.Builder<String, Object>()
+                .put(Message.FIELD_ID, "msg-id")
+                .put(Message.FIELD_TIMESTAMP, Tools.buildElasticSearchTimeFormat(Tools.nowUTC()))
+                .build());
+
+        msg.addProcessingError(new Message.ProcessingError(() -> "Cause", "Failure Message #1", "Failure Details #1"));
+
+        assertThat(msg.processingErrors()).hasSize(1);
+
+        assertThatCode(() -> msg.processingErrors().add(new Message.ProcessingError(() -> "Cause 2", "Failure Message #2", "Failure Details #2")))
+                .isInstanceOf(Exception.class);
+
+        assertThat(msg.processingErrors()).hasSize(1);
+    }
+
+    @Test
+    public void toElasticSearchObject_processingErrorDetailsAreJoinedInOneStringAndReturnedInProcessingErrorField() {
+        // given
+        final Message msg = new Message(new org.testcontainers.shaded.com.google.common.collect.ImmutableMap.Builder<String, Object>()
+                .put(Message.FIELD_ID, "msg-id")
+                .put(Message.FIELD_TIMESTAMP, Tools.buildElasticSearchTimeFormat(Tools.nowUTC()))
+                .build());
+
+        msg.addProcessingError(new Message.ProcessingError(
+                () -> "Cause 1", "Failure Message #1", "Failure Details #1"
+        ));
+
+        msg.addProcessingError(new Message.ProcessingError(
+                () -> "Cause 2", "Failure Message #2", "Failure Details #2"
+        ));
+
+        // when
+        final Map<String, Object> esObject = msg.toElasticSearchObject(new ObjectMapperProvider().get(), new Meter());
+
+        // then
+        assertThat(esObject.get(Message.FIELD_GL2_PROCESSING_ERROR))
+                .isEqualTo("Failure Details #1, Failure Details #2");
     }
 }
